+++ src/plone/restapi/deserializer/local_roles.py
@@ -47,7 +47,7 @@ class DeserializeFromJson(object):
         # roles
         roles_reindex = False
         new_roles = data.get("entries", None)
-        managed_roles = frozenset([r['id'] for r in sharing_view.roles()])
+        managed_roles = frozenset([r["id"] for r in sharing_view.roles()])
 
         if new_roles is not None:
             # the roles are converted into a FrozenSet so we have to filter