+++ src/plone/restapi/interfaces.py
@@ -98,6 +98,47 @@ class IFieldDeserializer(Interface):
         """
 
 
+class IBlockFieldDeserializationTransformer(Interface):
+    """Convert/adjust raw block deserialized value into block value.
+    """
+
+    block_type = Attribute(
+        "A string with the type of block, the @type from " "the block value"
+    )
+    order = Attribute(
+        "A number used in sorting value transformers. " "Smaller is executed first"
+    )
+
+    def __init__(field, context, request):
+        """Adapts context and the request.
+        """
+
+    def __call__(value):
+        """Convert the provided raw Python value to a block value.
+        """
+
+
+class IBlockFieldSerializationTransformer(Interface):
+    """Transform block value before final JSON serialization
+    """
+
+    block_type = Attribute(
+        "A string with the type of block, the @type from " "the block value"
+    )
+    order = Attribute(
+        "A number used in sorting value transformers for the "
+        "same block. Smaller is executed first"
+    )
+
+    def __init__(field, context, request):
+        """Adapts context and the request.
+        """
+
+    def __call__(value):
+        """Convert the provided raw Python value to a block value.
+        """
+
+
 class IExpandableElement(Interface):
     """A named adapter that deserializes an element in expanded or collapsed
        form.