+++ src/plone/restapi/serializer/blocks.py
@@ -2,6 +2,7 @@
 from plone.outputfilters.browser.resolveuid import uuidToObject
 from plone.outputfilters.browser.resolveuid import uuidToURL
 from plone.restapi.behaviors import IBlocks
+from plone.restapi.interfaces import IBlockFieldSerializationTransformer
 from plone.restapi.interfaces import IFieldSerializer
 from plone.restapi.interfaces import IObjectPrimaryFieldTarget
 from plone.restapi.serializer.converters import json_compatible
@@ -9,8 +10,10 @@ from plone.restapi.serializer.dxfields import DefaultFieldSerializer
 from plone.schema import IJSONField
 from zope.component import adapter
 from zope.component import queryMultiAdapter
+from zope.component import subscribers
 from zope.interface import implementer
 from zope.interface import Interface
+from zope.publisher.interfaces.browser import IBrowserRequest
 
 import copy
 import re
@@ -25,15 +28,40 @@ class BlocksJSONFieldSerializer(DefaultFieldSerializer):
     def __call__(self):
         value = copy.deepcopy(self.get_value())
 
-        # Resolve UID links
         if self.field.getName() == "blocks":
-            for block in value.values():
-                if block.get("@type") == "text":
-                    self.resolve_links(block)
+            for id, block_value in value.items():
+                block_type = block_value.get("@type", "")
+
+                handlers = [
+                    h
+                    for h in subscribers(
+                        (self.context, self.request),
+                        IBlockFieldSerializationTransformer,
+                    )
+                    if h.block_type == block_type
+                ]
+
+                for handler in sorted(handlers, key=lambda h: h.order):
+                    block_value = handler(block_value)
+
+                value[id] = block_value
+
         return json_compatible(value)
 
-    def resolve_links(self, block):
-        entity_map = block.get("text", {}).get("entityMap", {})
+
+@implementer(IBlockFieldSerializationTransformer)
+@adapter(IBlocks, IBrowserRequest)
+class TextBlockSerializer(object):
+    order = 100
+    block_type = "text"
+
+    def __init__(self, context, request):
+        self.context = context
+        self.request = request
+
+    def __call__(self, value):
+        # Resolve UID links
+        entity_map = value.get("text", {}).get("entityMap", {})
         for entity in entity_map.values():
             if entity.get("type") != "LINK":
                 continue
@@ -54,6 +82,7 @@ class BlocksJSONFieldSerializer(DefaultFieldSerializer):
                     entity["data"]["href"] = href
                     entity["data"]["url"] = href
                     print("SERIALIZE " + before + " -> " + href)  # noqa
+        return value
 
     def get_primary_field_target_url(self, uid):
         target_object = uuidToObject(uid)