+++ src/plone/restapi/tests/test_dxcontent_serializer.py
@@ -4,6 +4,7 @@ from datetime import datetime
 from datetime import time
 from datetime import timedelta
 from DateTime import DateTime
+from plone import api
 from plone.app.discussion.interfaces import IDiscussionSettings
 from plone.app.testing import logout
 from plone.app.testing import setRoles
@@ -88,8 +89,10 @@ class TestDXContentSerializer(unittest.TestCase):
         self.portal.doc1.modification_date = DateTime("2015-04-27T10:24:11+00:00")
         IMutableUUID(self.portal.doc1).set("30314724b77a4ec0abbad03d262837aa")
 
-    def serialize(self):
-        serializer = getMultiAdapter((self.portal.doc1, self.request), ISerializeToJson)
+    def serialize(self, obj=None):
+        if obj is None:
+            obj = self.portal.doc1
+        serializer = getMultiAdapter((obj, self.request), ISerializeToJson)
         return serializer()
 
     def test_serializer_returns_json_serializeable_object(self):
@@ -185,6 +188,224 @@ class TestDXContentSerializer(unittest.TestCase):
         obj = serializer()
         self.assertIn("is_folderish", obj)
         self.assertEqual(True, obj["is_folderish"])
+
+    def test_nextprev_no_nextprev(self):
+        folder = api.content.create(
+            container=self.portal,
+            type="Folder",
+            title="Folder with items",
+            description="This is a folder with some documents",
+        )
+        doc = api.content.create(
+            container=folder,
+            type="Document",
+            title="Item 1",
+            description="One item alone in the folder",
+        )
+        data = self.serialize(doc)
+        self.assertEqual({}, data["previous_item"])
+        self.assertEqual({}, data["next_item"])
+
+    def test_nextprev_has_prev(self):
+        folder = api.content.create(
+            container=self.portal,
+            type="Folder",
+            title="Folder with items",
+            description="This is a folder with some documents",
+        )
+        api.content.create(
+            container=folder,
+            type="Document",
+            title="Item 1",
+            description="Previous item",
+        )
+        doc = api.content.create(
+            container=folder,
+            type="Document",
+            title="Item 2",
+            description="Current item",
+        )
+        data = self.serialize(doc)
+        self.assertEqual(
+            {
+                "@id": "http://nohost/plone/folder-with-items/item-1",
+                "@type": "Document",
+                "title": "Item 1",
+                "description": "Previous item",
+            },
+            data["previous_item"],
+        )
+        self.assertEqual({}, data["next_item"])
+
+    def test_nextprev_has_next(self):
+        folder = api.content.create(
+            container=self.portal,
+            type="Folder",
+            title="Folder with items",
+            description="This is a folder with some documents",
+        )
+        doc = api.content.create(
+            container=folder,
+            type="Document",
+            title="Item 1",
+            description="Current item",
+        )
+        api.content.create(
+            container=folder, type="Document", title="Item 2", description="Next item"
+        )
+        data = self.serialize(doc)
+        self.assertEqual({}, data["previous_item"])
+        self.assertEqual(
+            {
+                "@id": "http://nohost/plone/folder-with-items/item-2",
+                "@type": "Document",
+                "title": "Item 2",
+                "description": "Next item",
+            },
+            data["next_item"],
+        )
+
+    def test_nextprev_has_nextprev(self):
+        folder = api.content.create(
+            container=self.portal,
+            type="Folder",
+            title="Folder with items",
+            description="This is a folder with some documents",
+        )
+        api.content.create(
+            container=folder,
+            type="Document",
+            title="Item 1",
+            description="Previous item",
+        )
+        doc = api.content.create(
+            container=folder,
+            type="Document",
+            title="Item 2",
+            description="Current item",
+        )
+        api.content.create(
+            container=folder, type="Document", title="Item 3", description="Next item"
+        )
+        data = self.serialize(doc)
+        self.assertEqual(
+            {
+                "@id": "http://nohost/plone/folder-with-items/item-1",
+                "@type": "Document",
+                "title": "Item 1",
+                "description": "Previous item",
+            },
+            data["previous_item"],
+        )
+        self.assertEqual(
+            {
+                "@id": "http://nohost/plone/folder-with-items/item-3",
+                "@type": "Document",
+                "title": "Item 3",
+                "description": "Next item",
+            },
+            data["next_item"],
+        )
+
+    def test_nextprev_root_no_nextprev(self):
+        data = self.serialize()
+        self.assertEqual({}, data["previous_item"])
+        self.assertEqual({}, data["next_item"])
+
+    def test_nextprev_root_has_prev(self):
+        doc = api.content.create(
+            container=self.portal,
+            type="Document",
+            title="Item 2",
+            description="Current item",
+        )
+        data = self.serialize(doc)
+        self.assertEqual(
+            {
+                "@id": "http://nohost/plone/doc1",
+                "@type": "DXTestDocument",
+                "title": "",
+                "description": "",
+            },
+            data["previous_item"],
+        )
+        self.assertEqual({}, data["next_item"])
+
+    def test_nextprev_root_has_next(self):
+        api.content.create(
+            container=self.portal,
+            type="Document",
+            title="Item 2",
+            description="Next item",
+        )
+        data = self.serialize()
+        self.assertEqual({}, data["previous_item"])
+        self.assertEqual(
+            {
+                "@id": "http://nohost/plone/item-2",
+                "@type": "Document",
+                "title": "Item 2",
+                "description": "Next item",
+            },
+            data["next_item"],
+        )
+
+    def test_nextprev_root_has_nextprev(self):
+        api.content.create(
+            container=self.portal,
+            type="Document",
+            title="Item 1",
+            description="Previous item",
+        )
+        doc = api.content.create(
+            container=self.portal,
+            type="Document",
+            title="Item 2",
+            description="Current item",
+        )
+        api.content.create(
+            container=self.portal,
+            type="Document",
+            title="Item 3",
+            description="Next item",
+        )
+        data = self.serialize(doc)
+        self.assertEqual(
+            {
+                "@id": "http://nohost/plone/item-1",
+                "@type": "Document",
+                "title": "Item 1",
+                "description": "Previous item",
+            },
+            data["previous_item"],
+        )
+        self.assertEqual(
+            {
+                "@id": "http://nohost/plone/item-3",
+                "@type": "Document",
+                "title": "Item 3",
+                "description": "Next item",
+            },
+            data["next_item"],
+        )
+
+    def test_nextprev_unordered_folder(self):
+        folder = api.content.create(
+            container=self.portal,
+            type="Folder",
+            title="Folder with items",
+            description="This is a folder with some documents",
+        )
+        folder.setOrdering("unordered")
+        doc = api.content.create(
+            container=folder,
+            type="Document",
+            title="Item 1",
+            description="One item alone in the folder",
+        )
+        data = self.serialize(doc)
+        self.assertEqual({}, data["previous_item"])
+        self.assertEqual({}, data["next_item"])
 
     def test_richtext_serializer_context(self):
         """This checks if the context is passed in correctly.
@@ -289,7 +510,7 @@ class TestDXContentSerializer(unittest.TestCase):
         self.assertEqual(True, obj["allow_discussion"])
 
     def test_allow_discussion_fti_allows_allows_global_enabled_but_no_instance_allowed(
-        self
+        self,
     ):  # noqa
         self.portal.invokeFactory("Document", id=u"doc2")
         registry = queryUtility(IRegistry)
@@ -307,7 +528,7 @@ class TestDXContentSerializer(unittest.TestCase):
         self.assertEqual(False, obj["allow_discussion"])
 
     def test_allow_discussion_fti_allows_allows_global_enabled_but_no_instance_set(
-        self
+        self,
     ):  # noqa
         self.portal.invokeFactory("Document", id=u"doc2")
         registry = queryUtility(IRegistry)
@@ -324,7 +545,7 @@ class TestDXContentSerializer(unittest.TestCase):
         self.assertEqual(True, obj["allow_discussion"])
 
     def test_allow_discussion_fti_disallows_allows_global_enabled_but_instance_allowed(
-        self
+        self,
     ):  # noqa
         self.portal.invokeFactory("Document", id=u"doc2")
         registry = queryUtility(IRegistry)
@@ -342,7 +563,7 @@ class TestDXContentSerializer(unittest.TestCase):
         self.assertEqual(True, obj["allow_discussion"])
 
     def test_allow_discussion_global_enabled_but_instance_has_no_discussion_behavior(
-        self
+        self,
     ):  # noqa
         registry = queryUtility(IRegistry)
         settings = registry.forInterface(IDiscussionSettings, check=False)
@@ -370,9 +591,7 @@ class TestDXContentPrimaryFieldTargetUrl(unittest.TestCase):
         )
 
         self.portal.invokeFactory(
-            "DXTestDocument",
-            id=u"doc2",
-            test_primary_namedfile_field=None,
+            "DXTestDocument", id=u"doc2", test_primary_namedfile_field=None,
         )
 
         self.portal.doc1.creation_date = DateTime("2015-04-27T10:14:48+00:00")
@@ -389,7 +608,10 @@ class TestDXContentPrimaryFieldTargetUrl(unittest.TestCase):
         data = serializer()
         self.assertIn("targetUrl", data)
         download_url = u"/".join(
-            [self.portal.doc1.absolute_url(), u"@@download/test_primary_namedfile_field"]
+            [
+                self.portal.doc1.absolute_url(),
+                u"@@download/test_primary_namedfile_field",
+            ]
         )
         self.assertEqual(data["targetUrl"], download_url)