+++ src/plone/restapi/tests/test_services_sources.py
@@ -179,7 +179,7 @@ class TestSourcesEndpoint(unittest.TestCase):
         )
 
     def test_context_source(self):
-        self.doc.title = u'Foo Bar Baz'
+        self.doc.title = u"Foo Bar Baz"
         transaction.commit()
 
         response = self.api_session.get(
@@ -202,11 +202,12 @@ class TestSourcesEndpoint(unittest.TestCase):
         )
 
     def test_source_filtered_by_non_ascii_title(self):
-        self.doc.title = u'Bär'
+        self.doc.title = u"Bär"
         transaction.commit()
 
         response = self.api_session.get(
-            "%s/@sources/test_choice_with_context_source?title=b%%C3%%A4r" % self.doc.absolute_url()
+            "%s/@sources/test_choice_with_context_source?title=b%%C3%%A4r"
+            % self.doc.absolute_url()
         )
 
         self.assertEqual(response.status_code, 200)
@@ -215,9 +216,7 @@ class TestSourcesEndpoint(unittest.TestCase):
             {
                 u"@id": self.portal_url
                 + u"/testdoc/@sources/test_choice_with_context_source?title=b%C3%A4r",  # noqa
-                u"items": [
-                    {u'token': u'b=C3=A4r', u'title': u'B\xe4r'},
-                ],
+                u"items": [{u"token": u"b=C3=A4r", u"title": u"B\xe4r"}],
                 u"items_total": 1,
             },
         )