+++ src/plone/restapi/tests/test_services_users.py
@@ -439,10 +439,10 @@ class TestUsersEndpoint(unittest.TestCase):
     def test_update_portrait(self):
         payload = {
             "portrait": {
-                "filename": "image.png",
+                "filename": "image.gif",
                 "encoding": "base64",
                 "data": u"R0lGODlhAQABAIAAAP///wAAACwAAAAAAQABAAACAkQBADs=",
-                "content-type": "image/png",
+                "content-type": "image/gif",
             }
         }
         self.api_session.auth = ("noam", "password")
@@ -459,10 +459,10 @@ class TestUsersEndpoint(unittest.TestCase):
     def test_update_portrait_with_default_plone_scaling(self):
         payload = {
             "portrait": {
-                "filename": "image.png",
+                "filename": "image.gif",
                 "encoding": "base64",
                 "data": u"R0lGODlhAQABAIAAAP///wAAACwAAAAAAQABAAACAkQBADs=",
-                "content-type": "image/png",
+                "content-type": "image/gif",
                 "scale": True,
             }
         }
@@ -480,12 +480,71 @@ class TestUsersEndpoint(unittest.TestCase):
     def test_update_portrait_by_manager(self):
         payload = {
             "portrait": {
-                "filename": "image.png",
+                "filename": "image.gif",
                 "encoding": "base64",
                 "data": u"R0lGODlhAQABAIAAAP///wAAACwAAAAAAQABAAACAkQBADs=",
-                "content-type": "image/png",
+                "content-type": "image/gif",
             }
         }
+        response = self.api_session.patch("/@users/noam", json=payload)
+
+        self.assertEqual(response.status_code, 204)
+        transaction.commit()
+
+        user = self.api_session.get("/@users/noam").json()
+        self.assertTrue(
+            user.get("portrait").endswith("plone/portal_memberdata/portraits/noam")
+        )
+
+    def test_delete_portrait(self):
+        payload = {
+            "portrait": None,
+        }
+        self.api_session.auth = ("noam", "password")
+        response = self.api_session.patch("/@users/noam", json=payload)
+
+        self.assertEqual(response.status_code, 204)
+        transaction.commit()
+
+        user = self.api_session.get("/@users/noam").json()
+
+        self.assertTrue(user.get("portrait") is None)
+
+    def test_delete_portrait_by_manager(self):
+        payload = {
+            "portrait": None,
+        }
+        response = self.api_session.patch("/@users/noam", json=payload)
+
+        self.assertEqual(response.status_code, 204)
+        transaction.commit()
+
+        user = self.api_session.get("/@users/noam").json()
+
+        self.assertTrue(user.get("portrait") is None)
+
+    def test_update_user_with_portrait_set_without_updating_portrait(self):
+        payload = {
+            "portrait": {
+                "filename": "image.gif",
+                "encoding": "base64",
+                "data": u"R0lGODlhAQABAIAAAP///wAAACwAAAAAAQABAAACAkQBADs=",
+                "content-type": "image/gif",
+            }
+        }
+        self.api_session.auth = ("noam", "password")
+        response = self.api_session.patch("/@users/noam", json=payload)
+
+        self.assertEqual(response.status_code, 204)
+        transaction.commit()
+
+        payload = {
+            "fullname": "Noam A. Chomsky",
+            "username": "noam",
+            "email": "avram.chomsky@plone.org",
+            "portrait": "http://localhost:55001/plone/portal_memberdata/portraits/noam",
+        }
+        self.api_session.auth = ("noam", "password")
         response = self.api_session.patch("/@users/noam", json=payload)
 
         self.assertEqual(response.status_code, 204)